Extract audio engine pill widget
