+++ src/gui_app/audio_settings.rs
@@ -1,20 +1,13 @@
-use radiant::gui::types::{Point, Rect, Rgba8};
-use radiant::layout::{LayoutOutput, Vector2};
 use radiant::prelude as ui;
-use radiant::runtime::{
-    PaintFillRect, PaintPrimitive, PaintStrokeRect, PaintText, PaintTextAlign, PaintTextRun,
-};
-use radiant::theme::ThemeTokens;
-use radiant::widgets::{
-    FocusBehavior, PaintBounds, PointerButton, TextWrap, Widget, WidgetCommon, WidgetInput,
-    WidgetOutput, WidgetSizing,
-};
 
 use super::{
     AUDIO_ENGINE_PILL_HEIGHT, AUDIO_ENGINE_PILL_ID, AUDIO_ENGINE_PILL_WIDTH,
     AUDIO_SETTINGS_POPUP_HEIGHT, AUDIO_SETTINGS_POPUP_WIDTH, GuiAppState, GuiMessage,
     VOLUME_SLIDER_HEIGHT, VOLUME_SLIDER_ID, VOLUME_SLIDER_WIDTH,
 };
+
+mod audio_engine_pill;
+pub(super) use audio_engine_pill::AudioEnginePill;
 
 pub(super) fn top_status_bar(state: &GuiAppState) -> ui::View<GuiMessage> {
     ui::row([
@@ -266,177 +259,5 @@ pub(super) fn format_sample_rate_label(sample_rate: u32) -> String {
         format!("{:.1} kHz", sample_rate as f32 / 1000.0)
     } else {
         format!("{sample_rate} Hz")
-    }
-}
-
-#[derive(Clone, Debug)]
-pub(super) struct AudioEnginePill {
-    common: WidgetCommon,
-    label: String,
-}
-
-impl AudioEnginePill {
-    pub(super) fn new(label: String, active: bool) -> Self {
-        let mut common = WidgetCommon::new(
-            0,
-            WidgetSizing::fixed(Vector2::new(
-                AUDIO_ENGINE_PILL_WIDTH,
-                AUDIO_ENGINE_PILL_HEIGHT,
-            )),
-        );
-        common.focus = FocusBehavior::Keyboard;
-        common.paint.bounds = PaintBounds::ClipToRect;
-        common.paint.paints_state_layers = false;
-        common.state.active = active;
-        Self { common, label }
-    }
-}
-
-impl Widget for AudioEnginePill {
-    fn common(&self) -> &WidgetCommon {
-        &self.common
-    }
-
-    fn common_mut(&mut self) -> &mut WidgetCommon {
-        &mut self.common
-    }
-
-    fn handle_input(&mut self, bounds: Rect, input: WidgetInput) -> Option<WidgetOutput> {
-        match input {
-            WidgetInput::PointerMove { position } => {
-                self.common.state.hovered = bounds.contains(position);
-                None
-            }
-            WidgetInput::PointerPress {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } if bounds.contains(position) => {
-                self.common.state.hovered = true;
-                self.common.state.pressed = true;
-                self.common.state.focused = true;
-                None
-            }
-            WidgetInput::PointerRelease {
-                position,
-                button: PointerButton::Primary,
-                ..
-            } => {
-                let activated = self.common.state.pressed && bounds.contains(position);
-                self.common.state.pressed = false;
-                self.common.state.hovered = bounds.contains(position);
-                activated.then(|| WidgetOutput::typed(GuiMessage::ToggleAudioSettings))
-            }
-            WidgetInput::FocusChanged(focused) => {
-                self.common.state.focused = focused;
-                if !focused {
-                    self.common.state.pressed = false;
-                }
-                None
-            }
-            WidgetInput::KeyPress(key) if self.common.state.focused => match key {
-                radiant::widgets::WidgetKey::Enter | radiant::widgets::WidgetKey::Space => {
-                    Some(WidgetOutput::typed(GuiMessage::ToggleAudioSettings))
-                }
-                _ => None,
-            },
-            _ => {
-                if matches!(input, WidgetInput::PointerRelease { .. }) {
-                    self.common.state.pressed = false;
-                }
-                None
-            }
-        }
-    }
-
-    fn accepts_pointer_move(&self) -> bool {
-        true
-    }
-
-    fn append_paint(
-        &self,
-        primitives: &mut Vec<PaintPrimitive>,
-        bounds: Rect,
-        _layout: &LayoutOutput,
-        _theme: &ThemeTokens,
-    ) {
-        let hovered_or_pressed = self.common.state.hovered || self.common.state.pressed;
-        let fill = if self.common.state.active {
-            Rgba8 {
-                r: 50,
-                g: 54,
-                b: 56,
-                a: 245,
-            }
-        } else if hovered_or_pressed {
-            Rgba8 {
-                r: 42,
-                g: 43,
-                b: 44,
-                a: 245,
-            }
-        } else {
-            Rgba8 {
-                r: 31,
-                g: 32,
-                b: 33,
-                a: 235,
-            }
-        };
-        primitives.push(PaintPrimitive::FillRect(PaintFillRect {
-            widget_id: self.common.id,
-            rect: bounds,
-            color: fill,
-        }));
-        primitives.push(PaintPrimitive::StrokeRect(PaintStrokeRect {
-            widget_id: self.common.id,
-            rect: Rect::from_min_max(
-                Point::new(bounds.min.x + 0.5, bounds.min.y + 0.5),
-                Point::new(bounds.max.x - 0.5, bounds.max.y - 0.5),
-            ),
-            color: Rgba8 {
-                r: 78,
-                g: 79,
-                b: 80,
-                a: if hovered_or_pressed { 230 } else { 165 },
-            },
-            width: 1.0,
-        }));
-        if self.common.state.focused {
-            primitives.push(PaintPrimitive::StrokeRect(PaintStrokeRect {
-                widget_id: self.common.id,
-                rect: Rect::from_min_max(
-                    Point::new(bounds.min.x - 1.0, bounds.min.y - 1.0),
-                    Point::new(bounds.max.x + 1.0, bounds.max.y + 1.0),
-                ),
-                color: Rgba8 {
-                    r: 255,
-                    g: 112,
-                    b: 86,
-                    a: 190,
-                },
-                width: 1.0,
-            }));
-        }
-        let font_size = 9.0;
-        let text_rect = Rect::from_min_max(
-            Point::new(bounds.min.x + 5.0, bounds.min.y),
-            Point::new(bounds.max.x - 5.0, bounds.max.y),
-        );
-        primitives.push(PaintPrimitive::Text(PaintTextRun {
-            widget_id: self.common.id,
-            text: PaintText::from(self.label.as_str()),
-            rect: text_rect,
-            font_size,
-            baseline: Some(((text_rect.height() - font_size) * 0.5 + font_size * 0.78).round()),
-            color: Rgba8 {
-                r: 183,
-                g: 184,
-                b: 184,
-                a: 235,
-            },
-            align: PaintTextAlign::Center,
-            wrap: TextWrap::None,
-        }));
     }
 }
